requirements.txt
fuzzywuzzy
aiohttp==3.6.0,<3.7.0
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
discord.py==1.3.2
idna==2.8
mojang-api-3==3.0.4
multidict==4.5.2
mysql-connector==2.2.9
python-box==3.4.0
requests==2.21.0
urllib3==1.24.3
websockets==6.0,!=7.0,!=8.0,!=8.0.1,<9.0
yarl==1.3.0
twython==3.7.0